+++ main.h
@@ -27,7 +27,7 @@ int add_str(va_list ap, char *buffer, int i);
 int add_str_print(va_list ap, char *buffer, int i);
 int copy_to_buff(char *s, char *buffer, int i);
 int add_percent(va_list ap, char *buffer, int i);
-char *int_to_str(unsigned long int num, int base);
+char *int_to_str(unsigned int num, int base);
 int add_int(va_list ap, char *buffer, int i);
 int add_unsigned_int(va_list ap, char *buffer, int i);
 int add_bin(va_list ap, char *buffer, int i);
